requirements.txt
# Core dependencies
anthropic>=0.30.0
fastapi>=0.115.6
uvicorn>=0.32.1
python-dotenv>=1.0.1
pydantic>=2.9.2
asyncio>=3.4.3
python-multipart >=0.0.19
transformers==4.47.0
tokenizers==0.21.0
huggingface-hub==0.24.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
mypy>=1.8.0
flake8>=7.0.0

# Type support
typing-extensions>=4.7.1

# SQLAlchemy
SQLAlchemy>=2.0.36
